+++ src/mono/mono/mini/interp/jiterpreter.c
@@ -381,13 +381,15 @@ mono_jiterp_conv (void *dest, void *src, int opcode) {
 	return 0;
 }
 
-#define JITERP_CNE_UN_R4 (0xFFFF + 0)
-#define JITERP_CGE_UN_R4 (0xFFFF + 1)
-#define JITERP_CLE_UN_R4 (0xFFFF + 2)
-#define JITERP_CNE_UN_R8 (0xFFFF + 3)
-#define JITERP_CGE_UN_R8 (0xFFFF + 4)
-#define JITERP_CLE_UN_R8 (0xFFFF + 5)
-
+// keep in sync with jiterpreter-opcodes.ts
+enum {
+	JITERP_CNE_UN_R4 = (0xFFFF + 0),
+	JITERP_CGE_UN_R4,
+	JITERP_CLE_UN_R4,
+	JITERP_CNE_UN_R8,
+	JITERP_CGE_UN_R8,
+	JITERP_CLE_UN_R8,
+};
 
 #define JITERP_RELOP(opcode, type, op, noorder) \
 	case opcode: \
@@ -1108,25 +1110,29 @@ mono_jiterp_stelem_ref (
 	return 1;
 }
 
-#define JITERP_MEMBER_VT_INITIALIZED 0
-#define JITERP_MEMBER_ARRAY_DATA 1
-#define JITERP_MEMBER_STRING_LENGTH 2
-#define JITERP_MEMBER_STRING_DATA 3
-#define JITERP_MEMBER_IMETHOD 4
-#define JITERP_MEMBER_DATA_ITEMS 5
-#define JITERP_MEMBER_RMETHOD 6
-#define JITERP_MEMBER_SPAN_LENGTH 7
-#define JITERP_MEMBER_SPAN_DATA 8
-#define JITERP_MEMBER_ARRAY_LENGTH 9
-#define JITERP_MEMBER_BACKWARD_BRANCH_OFFSETS 10
-#define JITERP_MEMBER_BACKWARD_BRANCH_OFFSETS_COUNT 11
-#define JITERP_MEMBER_CLAUSE_DATA_OFFSETS 12
-#define JITERP_MEMBER_PARAMS_COUNT 13
-#define JITERP_MEMBER_VTABLE 14
-#define JITERP_MEMBER_VTABLE_KLASS 15
-#define JITERP_MEMBER_CLASS_RANK 16
-#define JITERP_MEMBER_CLASS_ELEMENT_CLASS 17
-#define JITERP_MEMBER_BOXED_VALUE_DATA 18
+// keep in sync with jiterpreter-enums.ts JiterpMember
+enum {
+	JITERP_MEMBER_VT_INITIALIZED = 0,
+	JITERP_MEMBER_ARRAY_DATA,
+	JITERP_MEMBER_STRING_LENGTH,
+	JITERP_MEMBER_STRING_DATA,
+	JITERP_MEMBER_IMETHOD,
+	JITERP_MEMBER_DATA_ITEMS,
+	JITERP_MEMBER_RMETHOD,
+	JITERP_MEMBER_SPAN_LENGTH,
+	JITERP_MEMBER_SPAN_DATA,
+	JITERP_MEMBER_ARRAY_LENGTH,
+	JITERP_MEMBER_BACKWARD_BRANCH_OFFSETS,
+	JITERP_MEMBER_BACKWARD_BRANCH_OFFSETS_COUNT,
+	JITERP_MEMBER_CLAUSE_DATA_OFFSETS,
+	JITERP_MEMBER_PARAMS_COUNT,
+	JITERP_MEMBER_VTABLE,
+	JITERP_MEMBER_VTABLE_KLASS,
+	JITERP_MEMBER_CLASS_RANK,
+	JITERP_MEMBER_CLASS_ELEMENT_CLASS,
+	JITERP_MEMBER_BOXED_VALUE_DATA
+};
+
 
 // we use these helpers at JIT time to figure out where to do memory loads and stores
 EMSCRIPTEN_KEEPALIVE size_t
@@ -1176,20 +1182,23 @@ mono_jiterp_get_member_offset (int member) {
 	}
 }
 
-#define JITERP_COUNTER_TRACE_CANDIDATES 0
-#define JITERP_COUNTER_TRACES_COMPILED 1
-#define JITERP_COUNTER_ENTRY_WRAPPERS_COMPILED 2
-#define JITERP_COUNTER_JIT_CALLS_COMPILED 3
-#define JITERP_COUNTER_DIRECT_JIT_CALLS_COMPILED 4
-#define JITERP_COUNTER_FAILURES 5
-#define JITERP_COUNTER_BYTES_GENERATED 6
-#define JITERP_COUNTER_NULL_CHECKS_ELIMINATED 7
-#define JITERP_COUNTER_NULL_CHECKS_FUSED 8
-#define JITERP_COUNTER_BACK_BRANCHES_EMITTED 9
-#define JITERP_COUNTER_BACK_BRANCHES_NOT_EMITTED 10
-#define JITERP_COUNTER_ELAPSED_GENERATION 11
-#define JITERP_COUNTER_ELAPSED_COMPILATION 12
-#define JITERP_COUNTER_MAX JITERP_COUNTER_ELAPSED_COMPILATION
+// keep in sync with jiterpreter-enums.ts JiterpCounter
+enum {
+	JITERP_COUNTER_TRACE_CANDIDATES = 0,
+	JITERP_COUNTER_TRACES_COMPILED,
+	JITERP_COUNTER_ENTRY_WRAPPERS_COMPILED,
+	JITERP_COUNTER_JIT_CALLS_COMPILED,
+	JITERP_COUNTER_DIRECT_JIT_CALLS_COMPILED,
+	JITERP_COUNTER_FAILURES,
+	JITERP_COUNTER_BYTES_GENERATED,
+	JITERP_COUNTER_NULL_CHECKS_ELIMINATED,
+	JITERP_COUNTER_NULL_CHECKS_FUSED,
+	JITERP_COUNTER_BACK_BRANCHES_EMITTED,
+	JITERP_COUNTER_BACK_BRANCHES_NOT_EMITTED,
+	JITERP_COUNTER_ELAPSED_GENERATION,
+	JITERP_COUNTER_ELAPSED_COMPILATION,
+	JITERP_COUNTER_MAX = JITERP_COUNTER_ELAPSED_COMPILATION
+};
 
 #define JITERP_COUNTER_UNIT 100
 
@@ -1222,10 +1231,13 @@ mono_jiterp_modify_counter (int counter, double delta) {
 	return ((double)actual_result) / JITERP_COUNTER_UNIT;
 }
 
-#define JITERP_NUMBER_MODE_U32 0
-#define JITERP_NUMBER_MODE_I32 1
-#define JITERP_NUMBER_MODE_F32 2
-#define JITERP_NUMBER_MODE_F64 3
+// keep in sync with jiterpreter-enums.ts JiterpNumberMode
+enum {
+	JITERP_NUMBER_MODE_U32 = 0,
+	JITERP_NUMBER_MODE_I32,
+	JITERP_NUMBER_MODE_F32,
+	JITERP_NUMBER_MODE_F64,
+};
 
 EMSCRIPTEN_KEEPALIVE void
 mono_jiterp_write_number_unaligned (void *dest, double value, int mode) {
